ath10k: Support configuring firmware rate-ctrl-object count.
The firmware tries to cache rate-ctrl objects in the host (driver) memory.  But,
with lots of active stations, this appears to cause constant cache swapping and
in the end, rate-ctrl fails to work well at all.

CT Firmware has lots of RAM savings, especially when using fewer than 64 vdevs,
so allow users to configure more than the default (currently 32) of in-ram
rate-ctrl objects.  As long as firmware RAM is available, allocating as many
rate-ctrl objects as possible (up to number of peers) is probably a good idea.

This setting should be harmless in all CT firmware, but it will only have an
effect in beta-16 firmware and later.  Setting value to 0 means use firmware
defaults.

Signed-off-by: Ben Greear <greearb@candelatech.com>