ath10k:  Start supporting customization for 10.4 CT firmware.
This adds the framework to take the vdev-count, peer-count,
and other config from module-params and configure the 10.4
firmware.  Some of this will only work with CT 10.4 firmware.

Also, if we get WMI timeouts, pretend the firmware has crashed,
attempt to gather logs, and then restart the firmware.  Hopefully
the bug I found in 10.4 is now fixed in the CT 10.4 firmware
and this code will not actually be needed anymore!

Signed-off-by: Ben Greear <greearb@candelatech.com>